refactor: offset is now a long to avoid narrowing cast

+++ zeebe/snapshot/src/main/java/io/camunda/zeebe/snapshots/impl/SnapshotChunkUtil.java
@@ -31,7 +31,7 @@ final class SnapshotChunkUtil {
       final long snapshotChecksum,
       final String fileName,
       final byte[] fileData,
-      final int filePart) {
+      final long filePart) {
 
     final long checksum = createChecksum(fileData);
     return new SnapshotChunkImpl(
